bad type

+++ esphome/components/ota_http/ota_http_arduino.cpp
@@ -25,7 +25,7 @@ struct Header {
 int OtaHttpArduino::http_init() {
   int http_code;
   uint32_t start_time;
-  uint32_t long duration;
+  uint32_t duration;
 
   const char *header_keys[] = {"Content-Length", "Content-Type"};
   const size_t header_count = sizeof(header_keys) / sizeof(header_keys[0]);
